+++ changelog.md
@@ -6,6 +6,7 @@
 - Change villager trades for revolver pieces and mineral deposit maps to re-roll (BluSunrize)
     - After buying the item, the trade will be replaced with a new one
     - This only works on villagers acquired AFTER this update!
+- Prevent crates from being opened after being sealed (BluSunrize)
 - Fix villagers being unable to interact with IE doors (BluSunrize)
 - Fix steel doors dropping twice (BluSunrize)
 - Fix blueprint shelves not dropping blueprints when broken (BluSunrize)